requirements.txt
python-engineio==3.13.2 
python-socketio==4.6.1 
Flask-SocketIO==4.3.1 
eventlet
numpy
pillow
opencv-python
cryptography
python-statemachine
ipython
jupyter
ultralytics
transformers
simple-pid
sklearn
